requirements.txt
torch==1.9.0+cu111
torchvision==0.10.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.0.3
pyyaml==6.0.1
fvcore==0.1.5.post20221221
iopath==0.1.10